+++ ui/src/components/App.scss
@@ -1,4 +1,4 @@
-@import './common/CapTable.scss';
+@import './common/StripedTable.scss';
 @import './common/CardTable.scss';
 @import './common/DonutChart.scss';
 @import './common/FormErrorHandled.scss';
@@ -13,7 +13,6 @@
 @import './common/PageSection.scss';
 @import './common/PaginationControls.scss';
 @import './common/Profile.scss';
-@import './common/StripedTable.scss';
 @import './common/TabViewer.scss';
 @import './common/TopMenu.scss';
 @import './common/Wallet.scss';
@@ -36,7 +35,7 @@
         width: 100vw;
 
         /* common css */
-        @include cap-table;
+        @include striped-table;
         @include card-table;
         @include donut-chart;
         @include exchange;
@@ -53,7 +52,6 @@
         @include page-section;
         @include pagination-controls;
         @include profile;
-        @include striped-table;
         @include tab-viewer;
         @include top-menu;
         @include wallet;
