+++ ui/src/components/App.tsx
@@ -72,7 +72,7 @@ const PublicProvider: React.FC = ({ children }) => {
   const { parties, loading } = useDablParties();
   const { party, ledgerId, token } = computeCredentials(parties.publicParty);
 
-  return loading ? <OnboardingTile>Loading...</OnboardingTile> : (
+  return loading ? <OnboardingTile><p>Loading...</p></OnboardingTile> : (
     <PublicLedger
       ledgerId={ledgerId}
       publicParty={party}
