+++ ui/src/components/Custodian/ClientHoldings.tsx
@@ -11,7 +11,7 @@ import { RegisteredInvestor } from '@daml.js/da-marketplace/lib/Marketplace/Regi
 import { makeContractInfo } from '../common/damlTypes'
 import Page from '../common/Page'
 import PageSection from '../common/PageSection'
-import CapTable from '../common/CapTable';
+import StripedTable from '../common/StripedTable';
 import { depositSummary } from '../common/utils';
 
 import CreateDeposit from './CreateDeposit';
@@ -45,7 +45,7 @@ const ClientHoldings: React.FC<Props> = ({ sideNav, onLogout }) => {
             <PageSection className='clients'>
                 <div className='client-list'>
                     <Header as='h3'>Client Holdings</Header>
-                    <CapTable
+                    <StripedTable
                         headings={['Asset', 'Amount']}
                         rows={tableRows}/>
                 </div>
