+++ ui/src/components/Investor/Investor.tsx
@@ -148,7 +148,7 @@ const Investor: React.FC<Props> = ({ onLogout }) => {
         </InviteAcceptTile>
     );
 
-    const loadingScreen = <OnboardingTile>Loading...</OnboardingTile>
+    const loadingScreen = <OnboardingTile><p>Loading...</p></OnboardingTile>
 
     const investorScreen = <Switch>
         <Route exact path={path}>
