+++ ui/src/components/Issuer/Issuer.tsx
@@ -157,7 +157,7 @@ const Issuer: React.FC<Props> = ({ onLogout }) => {
         </InviteAcceptTile>
     );
 
-    const loadingScreen = <OnboardingTile>Loading...</OnboardingTile>
+    const loadingScreen = <OnboardingTile><p>Loading...</p></OnboardingTile>
 
     const sideNav = <RoleSideNav url={url}
                         name={registeredIssuer.contracts[0]?.payload.name || issuer}
