+++ ui/src/components/common/StripedTable.scss
@@ -1,13 +1,17 @@
+
 @mixin striped-table {
     .striped-table {
-        border: none;
+        display: flex;
+        flex-direction: column;
+        width: 100%;
+        align-items: flex-start;
 
-        table.striped-table {
+        table {
             border-collapse: collapse;
             border: none;
             margin: 0px;
 
-            .striped-table-header {
+            th {
                 border: none;
                 background-color: var(--cool-grey-100);
             }
@@ -16,31 +20,21 @@
                 border-top: 0px !important;
             }
 
-            .striped-table-body {
+            tbody {
                 background-color: var(--white);
                 border: 1px solid var(--cool-grey-90);
-                tr {
-                    background-color: red;
-                }
-                tr:nth-child(odd) {
-                    background-color: var(--blue-100);
-                }
             }
 
+            tr:nth-child(odd) {
+                background-color: var(--blue-100);
+            }
         }
 
-        // .striped-table-header > tr > th {
-        //     border: none;
-        //     background: none;
-        // }
 
-        // .striped-table-body > tr > td {
-        //     border: none;
-        //     background: none;
-        // }
-
-        // .striped-table-body > tr:nth-child(odd) {
-        //     background-color: var(--grey-3);
-        // }
+        @media screen and (max-width: 1000px) {
+            .ui.table.issuer-striped-table tr:nth-child(odd) {
+                background-color: unset;
+            }
+        }
     }
 }
